# laboratory_work-6/api/codegen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// GetForecastCityParams defines parameters for GetForecastCity.
type GetForecastCityParams struct {
	// City City name
	City []string `form:"city" json:"city"`
}

// GetForecastCoordinateParams defines parameters for GetForecastCoordinate.
type GetForecastCoordinateParams struct {
	// Coord List of coordinates in format 'lat,lon'.
	Coord []string `form:"coord" json:"coord"`
}

// GetWeatherCityParams defines parameters for GetWeatherCity.
type GetWeatherCityParams struct {
	// City City name
	City []string `form:"city" json:"city"`
}

// GetWeatherCoordinateParams defines parameters for GetWeatherCoordinate.
type GetWeatherCoordinateParams struct {
	// Coord List of coordinates in format 'lat,lon'.
	Coord []string `form:"coord" json:"coord"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get Forecast
	// (GET /forecast/city)
	GetForecastCity(c *gin.Context, params GetForecastCityParams)
	// Get Forecast
	// (GET /forecast/coordinate)
	GetForecastCoordinate(c *gin.Context, params GetForecastCoordinateParams)
	// Get Current Weather
	// (GET /weather/city)
	GetWeatherCity(c *gin.Context, params GetWeatherCityParams)
	// Get Current Weather
	// (GET /weather/coordinate)
	GetWeatherCoordinate(c *gin.Context, params GetWeatherCoordinateParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetForecastCity operation middleware
func (siw *ServerInterfaceWrapper) GetForecastCity(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetForecastCityParams

	// ------------- Required query parameter "city" -------------

	if paramValue := c.Query("city"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument city is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "city", c.Request.URL.Query(), &params.City, runtime.BindQueryParameterOptions{Type: "array", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter city: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetForecastCity(c, params)
}

// GetForecastCoordinate operation middleware
func (siw *ServerInterfaceWrapper) GetForecastCoordinate(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetForecastCoordinateParams

	// ------------- Required query parameter "coord" -------------

	if paramValue := c.Query("coord"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument coord is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "coord", c.Request.URL.Query(), &params.Coord, runtime.BindQueryParameterOptions{Type: "array", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter coord: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetForecastCoordinate(c, params)
}

// GetWeatherCity operation middleware
func (siw *ServerInterfaceWrapper) GetWeatherCity(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetWeatherCityParams

	// ------------- Required query parameter "city" -------------

	if paramValue := c.Query("city"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument city is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "city", c.Request.URL.Query(), &params.City, runtime.BindQueryParameterOptions{Type: "array", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter city: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetWeatherCity(c, params)
}

// GetWeatherCoordinate operation middleware
func (siw *ServerInterfaceWrapper) GetWeatherCoordinate(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetWeatherCoordinateParams

	// ------------- Required query parameter "coord" -------------

	if paramValue := c.Query("coord"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument coord is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "coord", c.Request.URL.Query(), &params.Coord, runtime.BindQueryParameterOptions{Type: "array", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter coord: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetWeatherCoordinate(c, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/forecast/city", wrapper.GetForecastCity)
	router.GET(options.BaseURL+"/forecast/coordinate", wrapper.GetForecastCoordinate)
	router.GET(options.BaseURL+"/weather/city", wrapper.GetWeatherCity)
	router.GET(options.BaseURL+"/weather/coordinate", wrapper.GetWeatherCoordinate)
}